pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "papers"
version = "0.1.0"
description = "Infraestrutura de coleta e análise para pesquisa empírica em GEO"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Alexandre Caramaschi", email = "caramaschiai@caramaschiai.io"},
]

dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "openai>=1.30",
    "anthropic>=0.34",
    "google-genai>=1.0",
    "supabase>=2.0",
    "python-dotenv>=1.0",
    "scipy>=1.13",
    "statsmodels>=0.14",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "pandas>=2.2",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
papers = "src.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
